refactor: create char and fc controllers
