+++ package.json
@@ -5,12 +5,12 @@
   "license": "MIT",
   "typings": "types/index.d.ts",
   "scripts": {
+    "build": "tsc",
     "express:start": "rimraf dist && tsc -p tsconfig.json && node dist/lib/server.js",
     "copyfiles": "copyfiles package.json readme.md dist",
     "release": "rimraf dist && tsc -p tsconfig.json && yarn run copyfiles && npm publish ./dist"
   },
   "dependencies": {
-    "@types/express": "^4.17.13",
     "axios": "^0.21.0",
     "express": "^4.17.1",
     "linkedom": "^0.13.0",
@@ -19,6 +19,7 @@
     "typescript": "^4.4.4"
   },
   "devDependencies": {
+    "@types/express": "^4.17.13",
     "@types/jsdom": "^16.2.13",
     "@types/lodash": "^4.14.176",
     "@types/node": "^16.11.6",
